+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/rest/request/refdata/v1/SearchSymbolRequestBuilderTest.java
@@ -1,0 +1,33 @@
+package pl.zankowski.iextrading4j.client.rest.request.refdata.v1;
+
+import org.junit.Test;
+import pl.zankowski.iextrading4j.api.refdata.v1.SymbolDescription;
+import pl.zankowski.iextrading4j.client.rest.manager.MethodType;
+import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
+
+import javax.ws.rs.core.GenericType;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.entry;
+
+public class SearchSymbolRequestBuilderTest {
+
+    @Test
+    public void shouldSuccessfullyCreateRequest() {
+        final String fragment = "apple";
+
+        final RestRequest<List<SymbolDescription>> request = new SearchSymbolRequestBuilder()
+                .withFragment(fragment)
+                .build();
+
+        assertThat(request.getMethodType()).isEqualTo(MethodType.GET);
+        assertThat(request.getPath()).isEqualTo("/search/{fragment}");
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<SymbolDescription>>() {
+        });
+        assertThat(request.getPathParams()).containsExactly(entry("fragment", fragment));
+        assertThat(request.getQueryParams()).isEmpty();
+    }
+
+
+}